Wrap breadcrumbs into a local domain and a propagated domain
Summary:
# Problem
We might want to change where we store "local" breadcrumbs, i.e breadcrumbs that were inferred during the current function. We could get more precise breadcrumbs by moving them all the way down in the taint representation.

The problem is that doing that is currently painful, because we make assumptions about the taint structure in many places. More specifically, we use `LocalTaint.update` with specific slots.

This is because manipulating breadcrumbs with the generic `transform BreadcrumbSet.Self ..` operator doesn't work: indeed, `BreadcrumbSet.Self` can refer to both the local breadcrumbs and the propagated ones, hence the LocalTaint abstract domain doesn't known which one it is.

# Solution
Let's introduce a new wrapper domain, which is a copy of an abstract domain with an additional `Self` part.
This allows us to use `transform LocalBreadcrumbSet.Self` and `transform PropagatedBreadcrumbSet.Self`.

Reviewed By: alexkassil

Differential Revision: D58363492

fbshipit-source-id: 5c79530da7cb35ac580aaa66d99efb9e17daf70c