Remove `GlobalResolution.less_or_equal_either_way`
Summary:
This function has exactly one module using it now, in `TypeInfo`'s
`meet` and `join`. It's nice to just reduce the surface area
by inlining; I was confused why this function even existed which
let to me simplifying the `TypeCheck` behavior and I think part
of the issue is that it shouldn't be in the GlobalResolution interface
in the first place.

(Note: I'm not really sure it's needed in `TypeInfo` either - I think
it's just a perf optimization that probably doesn't gain us anything
on realistic codebases, so I'll try removing it in a stacked diff. But
this diff is a pure refactor with no functional change.)

Reviewed By: grievejia

Differential Revision: D58424916

fbshipit-source-id: d52a5a2b72364fab25fe94a589c866ec6ae7ba28