Simiplify TypeInfo lattice operations
Summary:
We were short-circuiting attribute map join and meet in TypeInfo
join and meet when the base types weren't linearly orderable.

But I don't think this is really a good idea. It makes the code
more complex, and in principle there's no reason we can't
join or meet the attributes themselves - for example if I have
two incompatble types for `x`, `A` and `B` in two branches of an `if`
and I've refined `x.y` to be an `int` in both branches (let's say
it is final to avoid confusion with the unrelated question of soundness
of temporary refinements), then after the `if` it is perfectly valid for me to:
- treat the joined base type as `A | B`
- continue to have `x.y` refined to be `int`

So I think this linear orderability check is really jsut a performance
optimization. But attribute refinements are't really that common in the
first place, and as a result I don't think we're optimizing much.

I'd like to see the impact of simplifying this to do the proper thing
of just using a normal abstract map meet / join:
- do the type errors on IG change meaningfully?
- does it show up as a measurable perf hit?

Reviewed By: grievejia

Differential Revision: D58424915

fbshipit-source-id: 0a3b98cea5e4ce795dc347bc65d4250459cfd3c8