Add a test that shows a type variable scoping issue
Summary:
The way we handle type variables is currently wrong.

Consider the example in the testcase. In our current implementation, we generate constraints on funciton calls (so not on a program level). The constraint in this particular program will look like this:

T >= T -> T

This is not quite right because the "T" on the  RHS  is not actually the same as the "T" on the LHS. Instead, we could think of the constraint as:

T1 >= T -> T

but our implementation does not currently generate constraints correctly for this case.

Reviewed By: yangdanny97

Differential Revision: D58392805

fbshipit-source-id: 4e1c06f846bd05c164e0df6dd6928327ccf30bad