Parse type variables and convert them into variable aliases
Summary:
The way we handle type variables currently is that we collect them from the program and represent them as Type Aliases. This is a problem for two reasons:

1- Type variables are not ambigious. They are a type. So there is no need to delay processing them using Type Aliases. Instead, we can directly represent them with Variable Aliases

2- Using Type Aliases unnessesairly complicates the code readibility. Instead, we should process type variables in the parse_declaration function together with the Variadic.Parameters and Variadic.Tuple cases.

Changes:

This diff is a no-op, but adds a parsing function to parse type vars, and it modifies parse_declaration to use that output. We currently don't do anything with it, but we will do so in the next diff.

Reviewed By: stroxler

Differential Revision: D58393802

fbshipit-source-id: 3dcf5092c765ee14163626e9f8d9687da0d2b5d9