Don't let TypeInfo depend on GlobalResolution
Summary:
The dependency here made it not possible to move `Annotation.t`
into the `TypeInfo` module.

Taking callbacks as input instead of `GlobalResolution.t` is an
easy fix, and it makes the code a little easier to read anyway
since we no longer have inconsistent signatures across the
`LocalOrGlobal` -> `Annotation` boundary.

Reviewed By: yangdanny97

Differential Revision: D58474668

fbshipit-source-id: 03cf440b2deb7f52cc6d6b0fbca8d934f49c7768