Rename / move / simplify strange join-forcing-union function
Summary:
The `TypeInfo.GlobalOrLocal.join_annotation` function is pretty
strange; for historical reasons it lived in `TypeInfo.LocalOrGlobal`
but it actually operates on `TypeInfo.Unit`.

I'm not really sure why it's useful to force a Union in join, it
probably kills our perf in some cases compared to a proper join. But
that's a problem for a future diff (I filed a task and commented), because
it breaks tests if I get rid of that behavior right now.

This diff at least accomplishes two things:
- get rid of the absurd overhead of creating `TypeInfo.GlobalOrLocal`
  values out of `TypeInfo.Unit`, joining them, and then just pulling
  the `Unit` values back out again - that's pure obfuscation costing
  us CPU cycles and heap allocations
- make the name much clearer

Reviewed By: yangdanny97

Differential Revision: D58490563

fbshipit-source-id: 6cb9767d26fd28994313dbe5f65c98286573a7bd