Move `Annotation.t` -> `TypeInfo.Unit.t`
Summary:
This commit moves the code from `annotation.ml` into `typeInfo.ml`.

The idea here is that "type info" is the "stuff" Pyre stores in a type
context (wrapped up alongside globals in a "resolution") in order to
perform type checking. This "stuff" is more than just a type because
correct type narrowing requires mutability information that isn't
part of the type per se.

As a result, `TypeInfo.Unit.t` represents a single "unit" of type
information corresponding to a particular namable expression (which
could be a global or local, or an attribute access chain on either
one). This in turn gets wrapped up into a `GlobalOrLocal.t` which combines
the base unit for a particular global or local variable with all
attribute chains, and then into a `Store.t` which is a map of all
the `GlobalOrLocal.t`s and is the actual lookup map that powers
`Resolution.t` behavior.

Putting all these in one place makes it a little easier to see how
things work, easier to envision improvements (the current approach
involves two entire trees, one with "permanent" info and the other
with "temporary" refinements; it's not clear to me this is the best),
and most improtantly it frees us from using the name `Annotation.t`
which wasn't a great name for the pair of a type + mutability info.

--------

Note: there's still a lot of cleanup to do for variable names, function
args, and record fields named `annotation`. I'll work on stacking
some cleanups next.

Reviewed By: yangdanny97

Differential Revision: D58474666

fbshipit-source-id: 7616c022f32534cc3d0fe531824009588424ddc4