Update callee information incrementally
Summary:
The callee table is updated when checking a function, so it happens in the
TypeEnvironment part of the environment stack, which is dependency tracked.
However, the callgraph information is just stored in a non-tracked table,
Callgraph.SharedMemory.

This results in two issues:
1. callee information is not updated when a function is updated, because "add"
is a noop when the key already exists
2. callee information for deleted functions is never removed, so queries
continue to return results even when the function doesn't exist

This diff moves the callee information into TypeCheck.CheckResult alongside
other type error information. This data is part of the dependency-tracked
TypeEnvironment, so it is updated incrementally.

This diff also removes the distinction between FunctionCaller and
PropertySetterCaller keys. The Callgraph.SharedMemory table distinguished
between these, but the TypeEnvironment table is just keyed on reference. We
could choose to store callees for both the getter and setter under the same key,
like we do for errors, but this diff chooses to ignore the setter callees. This
preserves the existing behavior of the "callees" query API, which never returned
setter callees.

Reviewed By: stroxler

Differential Revision: D58387518

fbshipit-source-id: 445599f871fb07a34ec0d324d5decbdac70c29a8