Rename annotation -> type_info in several places
Summary:
This commit is nowhere near a complete sweep of Pyre's codebase, but it
does rename a lot of the obvious places where we use `TypeInfo.Unit.t`
values so we call the value `type_info` instead of `annotation`.

And a number of the dangling places are easy to find and clean up later
since we can look for `type_info:annotation` to find where some more renaming
is probably in order.

Reviewed By: yangdanny97

Differential Revision: D58496045

fbshipit-source-id: 581838537d2869fc5933ba0b47ff1d11b16cc938