Additionally store locations of the "triggers" in `TriggeredSinkHashMap`
Summary:
So that later the locations can be used to avoid generating triggered sinks on
the locations where the triggers are at.

I also decided to improve readability by updating the value type of
`TriggeredSinkHashMap.t`. It used to be a map from triggering sources to the
extra traces, but this is not understandable without any comments. With this
diff, I updated it into `trigger`, which represents the information of the
"trigger". There should be no changes w.r.t. the space complexity, since we
still store the same information.

Reviewed By: arthaud

Differential Revision: D58592428

fbshipit-source-id: 9c6b21b253c0815d9832d1c677143e4163a61349