Get rid of weird `...` missingFromStubs
Summary:
I think at some point in the ancient history of Pyre, our manual
parser treated `...` as an *identifier* rather than a built-in
AST node.

That's the only way I can make any sense of the code that injects
the fake name `...` into the global environment. I'm pretty sure it's
impossible for this to have an effect because Pyre doesn't as far as I
know contain any mechanism to produce the unparsable name "..." as
an Identifier or Reference, we *always* handle ellipses as the special
AST node `Expression.Ellipses`.

Let's simiplify.

Reviewed By: yangdanny97

Differential Revision: D58651242

fbshipit-source-id: cedfab4472930032604f10750dbf577f62f3cc14