Rearrange missingFromStubs.ml a bit
Summary:
I am beginning to look at combining UnannotatedGlobal with
ClassSummary so that they both live in the same symbol table.

Part of that cutover will be that `missingFromStubs` should output
both missing classes and missing globals at the same time. This commit
just tweaks the codebase so that we can do this with a smaller diff,
and so that when we are doing both at the same time it's clearer
what the `make_class` function is doing.

Reviewed By: yangdanny97

Differential Revision: D58613683

fbshipit-source-id: f7741b0e345dd5986e893c539c0d50ad1ddd9dee