Move special-case `type` into classSummary.ml
Summary:
I want to simplify the logic in unannotatedGlobalEnvironment.ml
so that we can reduce duplication in the code; the environment
ought to be responsible only for the plumbing needed to store
cached values and handle depndency tracking and not much
else.

To this end, move the special casing of `type` into
`ClassSummary.create`.

Reviewed By: yangdanny97

Differential Revision: D58651240

fbshipit-source-id: 04e445cebe77f6c228e7b8edfca732d5ab276e60