Simplify UnannotatedGlobal structure and names
Summary:
I want to combine UnannotatedGlobal and Module into a single ocaml module that
we can use for all of the base-level "raw" signatures  that Pyre extracts in
unannotatedGlobalEnvironment (except for classSummary, which is big enough to
stay in its own module at least for now); we have some major issues with
duplication and inconsistency that are hard to solve as long as these are
separate.

But before I combine them, it makes sense to simplify them as much as possible
by improving names and getting rid of any nested modules that aren't really needed:
- the `define_signature` and `import` datatypes are really just types that exist
  for the sake of immediate use in `UnannotatedGlobal.t`; having modules for them
  isn't really needed or particularly helpful, so I just use local type defs. This
  will make it easier when I combine modules since I'll need to introduce more nesting.
- the names of some of the variants weren't as helpful / specific as I'd like, for
  example an `import from` statement produced a `Name` variant which doesn't seem
  as clear as an `ImportFrom` variant.

This commit is pure reorg and rename, no logical changes occur whatsoever.

Reviewed By: grievejia

Differential Revision: D58651237

fbshipit-source-id: 15d4eeb17fc699967d0cb88a3ecedf06e4998015