Move "deprecated" code to the only callsite
Summary:
It can be pretty confusing having 4+ year old apis marked as
deprecated. If we don't want this logic being used, given that there is
only one callsite today the single easiest way to avoid it is to hardcode the
use case - in typeAliasEnvironment - instead of defining an API that no one
should rely on.

Reviewed By: yangdanny97

Differential Revision: D58651239

fbshipit-source-id: 9312db9a19dbd9a0dffdff5a7153eeb08f0e60c4