Remove `missing_from_stubs_builtin_globals` by supporting `__debug__`
Summary:
The name `__debug__` is unassignable in Python, so it behaves as a constant
boolean (it is fixed by the optimization level per interpreter). As a result,
we should treat it as a final bool value.

By doing this directly in `typeCheck.ml` instead of mucking with the global
environment, we can both better support it (because now it allows not just
temporary refinements) and simplify our environment-loading code; the need
to inject stuff not in the stubs makes it much harder to refactor our
environment-loading which is how I came across this in the first place.

Reviewed By: grievejia

Differential Revision: D58651241

fbshipit-source-id: 3e9927ab787fae1a21e2b4bcfd80c56e010c8fad