Move the unannotated global collection to Module.ml
Summary:
Due to the way the code evolved over time, the logic for how
to process a source into data describing the signature has been
split across several modules for some time, with some of the logic
living directly in unannotatedGlobalEnvironment.

This isn't desirable, because it's very hard to deduplicate
computation (for example the moduel metadata and module components
independently both run the unannotated global collector) in a refactor
unless these all live in the same place.

So, this diff continues consolidating this logic into `module.ml`,
creating a `Module.Components` namespace that contains the logic
formerly in the `ModuleComponents` namespace of unannotatedGlobalEnvironment

Reviewed By: grievejia

Differential Revision: D58652273

fbshipit-source-id: 59946809519966c674da467a5fcaa747297a16e1