Switch Module.Components to using maps instead of alists.
Summary:
Once data is stored in unannotatedGlobalEnvironment, the ordering of
definitions is never used in a meaningful way; in addition, for names that
are defined more than once, only the first definition we process can actually
be stored (for now - we may decide to index definitions in order to unblock
certain uses of re-definitions in the future).

So, it is better to have the data format directly represent the semantics here
by using maps rather than alists.

In the process, you can see that some Pysa test behaviors actually change, where
they were relying on duplicated and in-order names. I think this is okay since
you can't really have multiple globals of the same name anway, although I will
ping the Pysa folks to make sure.

Reviewed By: tianhan0

Differential Revision: D58692966

fbshipit-source-id: eed6da1606a0cd032d507212554c4f3c835a0c24