Combine unannotatedGlobal.ml and module.ml
Summary:
Both unannotatedGlobal.ml and module.ml contain logic related
to "deterimining what is in a module", as in all the global names.

In hindsight, it was probably never a great idea to separate these two
components as much as we did; it has led to duplication where we compute
the UnannotatedGlobal list multiple times, post-processing it inconsistently,
in order to power the module metadata vs the UnannotatedGlobal.t table.

It will be much easier for me to simplify, deduplicate, and improve the
logic if we consolidate the logic in one module and also reduce the number
of external uses. This commit puts the data in one place; I will follow up
with a smaller diff moving some downstream logic that should really live
in the same place (the define folding in unannotatedGlobalEnvironment and
the wildcard name extraction in AstProcessing).

Reviewed By: grievejia

Differential Revision: D58651236

fbshipit-source-id: 3cb83e5f4e9104da9ec1ba3ba0c247356a3ebaee