Simplify `module.ml` logic by getting rid of `Collector`
Summary:
There's no real benefit to `Collector.Result.t list` over just using an
alist; the latter is actually *more* idiomatic for this sort of thing, and is
also consistent with what all the other module-level analysis data uses.

By getting rid of this and moving the actual collection into a helper
function in `UnannotatedGlobal` we also simplify the layout of the module
a bit which makes it easier to navigate.

Reviewed By: grievejia

Differential Revision: D58676102

fbshipit-source-id: c1fda9351016b79202bff8fde23e1ad68ef58477