Move `wildcard_exports_of_raw_source`
Summary:
I'm trying to collect all of the "collect things from a module"
logic in one place.

The `wildcard_exports_of_raw_source` logic is different from other use cases
because it involves using *raw* rather than *preprocessed* sources and is used
in preprocessing, but nontheless it relies on the same data and collection
logic, so it makes sense to keep it with the other use cases (particularly
when looking at major refactors).

Reviewed By: grievejia

Differential Revision: D58651238

fbshipit-source-id: bb550cbbfba35967db476bc3012f5c8968460c24