Warning messages about max number of overrides for beginners (#878)
Summary:
https://github.com/facebook/pyre-check/issues/870

This pull request addresses the issue of warnings related to classes with a high number of overrides in Pysa. It includes the following changes:

**Update Override Warning Messages:**

Enhanced the warning messages for classes with many overrides to include a suggestion to use the maximum_overrides_to_analyze option.
Added a link to the relevant documentation for further guidance.

**Configuration Updates:**

Added the maximum_overrides_to_analyze option to the taint.config files for different exercises to improve analysis performance and reduce warning messages.

**Please provide any feedback as comments for this pull request. I will then apply the feedback and update the code. **

Pull Request resolved: https://github.com/facebook/pyre-check/pull/878

Reviewed By: tianhan0

Differential Revision: D58673147

Pulled By: arthaud

fbshipit-source-id: 364a5ec6865c737562906cda6d164120e8698b05