Add kind-specific breadcrumbs for via:transform-tito-depth:N
Summary:
# Problem
The `via:transform-tito-depth:N` is kind-specific: it should only be attached to flows that went through the transform.
However, right now, breadcrumbs are stored in `LocalTaint`, which means that breadcrumbs for a given call info are all mixed together, regardless of taint kind. This means we attach `via:transform-tito-depth:N` to taint that didn't go through the transform, making it confusing.

# Solution
We could move local breadcrumbs down in the taint representation (in `Frame`), so that their are tied to a specific kind.
We have tried this and it lead to a performance regression of 6%.

Instead, what this diff does is have a special set for kind-specific breadcrumbs that are stored in `Frame`. This is then used for `via:transform-tito-depth:N`. This way, we get the right semantic, and preserve performance.

Reviewed By: tianhan0

Differential Revision: D58468250

fbshipit-source-id: 4889f78c5d57b211ec466ac7794fd0661751bb0f