Move subtraces down in the taint representation
Summary:
Right now, "extra_trace", which is the field holding the set of subtraces for a given taint, is stored in `LocalTaint`, which means it one set for all the different taint kinds.
In practice, we do want to have different traces for each kind. For instance, when we add a transform to a taint, we only want to add the subtrace to that transform flow, and not other taint.
To fix this, let's store subtraces in `Frame`, which is kind-specific.

Reviewed By: tianhan0

Differential Revision: D58465386

fbshipit-source-id: a894dacb74f16629a5d7fdc0a12a72e376a72e76