Convert some tests to use better names
Summary:
Using a manually-chosen prefix for test names works just fine
when iterating on a single test, but it turns out it causes problems
for big refactors that need us to `make test` because it's not always
obvious where a failure is even coming from - the module name only gets
printed once at the top of a string of failures, and if the failures are
big it can be hard to even scroll to the top.

This was really bad for some of the environment tests because they
were almost *all* using "environment" as the top-level name!

I already discovered with labeled_test_case that automating the
names is a simple but powerful way to get better ounit output;
let's just do the same with the overal test suite names via
`__MODULE__`. Dune injects junk into the names, so I use a helper
to strip that out.

This commit only changes a handful of tests, mostly the environment layers.

Reviewed By: yangdanny97

Differential Revision: D58847758

fbshipit-source-id: fda4a7e9ba4ed9f36b263219e7ebf250f28ec09d