Split update test cases in unannotatedGlobalEnvironmentTest
Summary:
It's hard to split all of the unannotatedGlobalEnvironmentTest unit tests
because it doesn't follow a set pattern the way integration tests do,
but the update tests make up a really big chunk of the overall tests
and they are pretty easy to split since each test case uses a fresh scratch
project.

This commit splits just those tests. The motivation is that I'm attempting
a nontrivial refactor of UGE, and hitting test failures that I need to be
able to identify and print-debug quickly, and without split test cases it
is not really possible to be efficient.

Reviewed By: yangdanny97

Differential Revision: D58852363

fbshipit-source-id: 793ccefc798e070c36776fc15e9c7eb7bf683461