do not emit "undefined type" error for ... annotation
Summary:
The current behavior is unintuitive, I think it's trying to look up `...` as a class and fails, giving a confusing error message. It only appears once per file for the conformance tests (and presumably in prod), so if a previous line has an invalid `...` then later occurrences will not emit an error until the first one is resolved.

I'd rather have clearer/more actionable error messages when we see ellipses in invalid places. Tuples have a similar problem which I'll address in another diff.

Reviewed By: grievejia

Differential Revision: D58888949

fbshipit-source-id: 4877d2c5b6a24a862bc56dcc144036d39e474bbb