Simplify `Type.TypedDictionary`, clarify its role
Summary:
In Pyre today, TypedDictionary is not a type.

While I think this is probably a bug in the type system (I'd really
like to eventually distinguish "raw" annotations from actual types,
and make types aware of things like TypedDictionary and Protocol),
it should be accurately reflected in the code for now.

This means `TypedDictionary` should not be part of the `Type.Record`
module, nor should its type be generic over `'annotation'` - instead,
it's just a simple derived type that depends on `Type.t` directly.

Reviewed By: grievejia

Differential Revision: D58976853

fbshipit-source-id: 063c559d070befeaa16c399a79740e36a34f8399