Use a custom type for callable parameter names
Summary:
There's some hard-to-simplify-cleanly logic right now in the
manually-written `equal` functions for callable types, which exists
only to deal with side effects of qualification: qualification will
rename function parameters to a locally-qualified name with a `$`
in it, and as a result we have to sanitize the names before we can
get semantically-correct-for-python "equal" logic.

This commit gets rid of the handwritten equal and instead uses
a custom type to handle this, and along the way adds a comment to
make the rationale more explicit (before you had to sort of reverse
engineer the "why").

Reviewed By: grievejia

Differential Revision: D58987227

fbshipit-source-id: d27e96df606d822438e7c23f65ad7f4f5df3be39