Don't use a weakened_type typed_dictionary.
Summary:
Jia and I would both like to make the `Type.t` type rely much less
on generics than it does today. A clear step forward would be to make
`Type.TypedDictionary.t` depend directly on `Type.t`.

The only place where the generic-ness is currently used is weakenMutableLiterals.
If we really thought it was a huge win, I'd be okay with keeping the generic
behavior. But the generic really isn't gaining us much here, all it does is avoid
using a tuple in one of the internal handoffs between two nested functions. That's
not a good reason to make `Type.ml` more confusing, it's better to have
the `weakenMutableLiterals` code be slightly more verbose than to muck up an
interface used throughout the Pyre codebase.

Reviewed By: grievejia

Differential Revision: D58976854

fbshipit-source-id: e3c81b9e83d44701e0d808b1750f66badad68b5b