fix concise error messages for some types
Summary:
when printing types in concise mode, the `strip_qualification` helper splits identifiers on `.` and takes the last part, turning `typing.Tuple` into `Tuple`, etc.

it's only intended to be used for identifiers, but there are some places in the codebase which render a raw type annotation (expression) as a string and pass it to the pretty printer as a primitive

for example:

https://www.internalfb.com/code/fbsource/[c3c752f598b062c75b6a6a3f892334e962fd42ae]/fbcode/tools/pyre/source/analysis/typeCheck.ml?lines=590

this means that `Callable[[...], object]` actually gets passed to the error message as `Primitive "Callable[[...], object]"`, which gets treated as an identifier and stripped down to `], object]`

this diff adds a new `InvalidTypeAnnotationExpression` variant to `invalid_type_kind` which takes an expression instead of a resolved type, and renames the original `InvalidType` to `InvalidTypeAnnotation`.

the expression can now be passed in directly without first turning it into a string and wrapping it with `Type.Primitive`, solving our problem.

there were two cases of enum-related errors which pass a type to this error, those remain on the old error format since they don't do this weird conversion.

two alternatives I thought of:
1. just pass the resolved type to the original error instead of stringifying the expression - this won't work for a lot of cases since malformed types can resolve to `unknown` and display as such, making the error message useless since it doesn't match what the user wrote
2. make `strip_qualification` skip identifiers with `...` - if `Type.Primitive (Expression.show expression)` is a hack, that feels like a hack on top of a hack

Reviewed By: migeed-z

Differential Revision: D59023970

fbshipit-source-id: e6e4a09283e588d5ae76224fae77dad7e693916f