Move OrderedTypes logic to OrderedTypes module
Summary:
In general, the type.ml code follows a pattern that nontrivial
logic should live in top-level modules defined *below* the core
data, which makes it much easier to skim the type definitions.

The only really major exception to this is that for whatever
reason, at some point we put some really huge utility functions
in `Record.OrderedTypes` instead of putting it in the
`Type.OrderedTypes` module.

I'm currently trying to make the type defs readable (and with
some luck maybe even simplify them a bit), and these functions
are very much in my way.

There's no reason not to put them in the conventional spot, so
this diff moves them.

Reviewed By: migeed-z

Differential Revision: D59018121

fbshipit-source-id: d98a7bdb2f8fe68bf96a22973d1be7caee3d4af4