Clean up dangling Type.Variable module / constructor names
Summary:
This gets rid of some dangling references to Unary, ParameterVariadic,
and TupleVariadic, converting them to TypeVar, ParamSpec, and TypeVarTuple.

Let's clearly distinguish the constructors for `Type.Variable.t`
(which is not itself a type actually, it's just a "package") used
to unify type var like things e.g. in `variable_aliases` callbacks)
from the modules of `Variable` by making the constructor names
explicitly have a postfix `Varaible` in the name.

Reviewed By: yangdanny97

Differential Revision: D59005126

fbshipit-source-id: 1c4f8a48a3ddd6ba3a98b94b023f3b7960f645c8