Rename ParameterVariadicComponent to ParamSpecComponent
Summary:
What this is is a "part" of a param spec like `P.args` and `P.kwargs`, which
can be used directly as an annotation (on a variadic parameter, so `*args` or
`*kwargs`) in a context where `P` is bound.

The name `ParamSpecComponent` seems much clearer to me, plus it's shorter.

Reviewed By: migeed-z

Differential Revision: D59026651

fbshipit-source-id: 8b4790cd5093953aa3917213d2061c8444a3e91e