Extract pretty printing logic from `Record`
Summary:
I am on a quest right now to make it easy to see the data layout
of `Type.t` through a combination of improving the names and getting
rid of the logic that's interspersed amongst type definitions.

A great deal of the logic causing me problems is actually coming from
pretty printing; many of the pretty printers are defined as helpers
inside of the `Record` module.

This commit moves all of those helpers to `PrettyPrinting`

In the process I also eliminated every function that accepts
`~pp_type` as an input from the public interface - we should only expose
pretty printers that know how they want to print type, because the fact
that type printing is injectable is an implementation detail used mainly
to do type gymnastics and to have both `pp` and `pp_concise`; it shouldn't
be leaking to callers.

Reviewed By: migeed-z

Differential Revision: D59018122

fbshipit-source-id: 17bd6889fbb5803e0bc0b36a9f8eb0ad25de6d89