Rename Callable.Parameter to CallableParamType
Summary:
This is probably my most controversial name change, but there are two
issues with `Callable.Parameter` as a name given the realities of Pyre code:

First, in a large fraction of the contexts where we want to `open Type.Callable`
we already have an `open Type`, but that means that `Parameter` is shadowed,
because `Type.Parameter` is a completely different concept from
`Type.Callable.Parameter`. This can easily create some confusion.

Second, the single most important user of `CallableParamType` is attributeResolution,
which converts the AST into the type-level representation. But in that code, it's
very common to have the raw AST `Parameter` module that lives inside `Ast.Statement`
in scope, so once again we can easily get name shadowing.

Name shadowing in itself is of course just a nuisance, but it also does affect how
easy it is to skim code, because it won't be quite as apparent at a glance whether,
for example, a specific line in attributeResolution is operating at the AST level
or the type level (I've run into difficulty skimming this code myself, e.g. when
working on the `__getitem__` refactors in May).

So disambiguating the names is helpful, and it's useful to use a name
that emphasizes both that this is tied to callables (which makes it
unlikely to be confused with `Type.Parameter` and that it is about
types (which makes it unlikely to be confused with the AST
`Parameter`). In the process we can also eliminate the
silly `RecordParameter` name, which wasn't really helping us any.

Reviewed By: grievejia

Differential Revision: D59021752

fbshipit-source-id: b920f40afacd91451d0825757cd90decbaaa9ee8