Move some code around
Summary:
I want to get all the pretty-printing helper functions out of
the `Record` module because they are one of the big blockers to
cleaning up the actual type definitions in `type.ml`.

To do that though, I have to tweak some orderings to avoid dependency
mismatches when I do the refactor.

This diff just moves some modules around in preparation for that work,
there's no actual change here.

Reviewed By: migeed-z

Differential Revision: D59018123

fbshipit-source-id: fe416ef9aae0b2920b09213b3b5377731e4c670c