Rename add to add_and_simplify in constraint generation
Summary:
Context:

Since we solve constraints greedily in our constraint generation process, calling the function that adds and solves constraints on the fly just `add` is misleading. Instead, the name should  indicate that we are also trying to solve.

If we look at the signature of the function, we can see that goes from constraint set to constraint set. Which means that we do not necessarily end up resolving every constraint in the set, which is why the new name indicates that.

Reviewed By: stroxler

Differential Revision: D59021457

fbshipit-source-id: d1b2facc6b2d3fb4b747186ad776336c34e29b8c