Fix the false negative that fails to propagate the triggered sinks on actual arguments at some call sites
Summary:
# Problem
For context, the triggered sinks are created by the forward analysis (which are
attached to actual arguments at a call site), but they need to be propagated by
the backward analysis. Hence, we store and pass the triggered sinks from the
forward analysis to the backward analysis.

Now, the problem is at the data structure (i.e., `TriggeredSinkLocationMap.t`)
that is used during the pass. Currently, we use a map from `AccessPath.Root` to
triggered sinks. For context, `AccessPath.Root` is used to represent the
variables that the taint analysis cares about -- such variables may have taint.
Using a map from `AccessPath.Root` to taints causes false negatives, because it
is possible that at a call site, the actual arguments can't be resolved into
`AccessPath.Root`. As a result, such taint can't be stored and passed to the
backward analysis.

# Solution 1 (This Diff)
The key change is the value type of `TriggeredSinkLocationMap.t`. It changed
from a `BackwardState` (i.e., a map from `AccessPath.Root`s to taint trees) to
a map from expressions to taint trees. This *directly* solves the false
negative when an expression can't be resolved into `AccessPath.Root`, because
the core problem is to pass taints (triggered sinks) on expressions from the
forward analysis to the backward analysis.

# Solution 2
An alternative solution is to map the actual arguments at call sites into
"fake" `AccessPath.Root`, so that any actual argument is guaranteed to be
resolved into an `AccessPath.Root`. Hence, all triggered sinks can now be
stored and passed to the backward analysis.

This approach also works, but
- it is a bit less direct than Solution 1, because the core problem is to pass
  taints on *expressions* (rather than `AccessPath.Root`, which is a subset of
  expressions) from the forward analysis to the backward analysis.
- it requires to maintain the extra code to map actual arguments to "fake"
  `AccessPath.Root`.

Reviewed By: arthaud

Differential Revision: D58822673

fbshipit-source-id: d79d0cc53f0568fd58fe024c49e3c43ed24a666f