Fix ppx issue with Response.Ok name collision
Summary:
I'm trying to upgrade ppx_deriving_yojson, but ran into an issue with the
Response.Ok constructor. This ppx generates code that returns a result, which
uses the Ok constructor as well.

This clash causes the generated code to fail to compile, because the two Ok
constructors get confused.

This diff works around the issue by renaming the constructor to Ok_ and using
the [name ...] annotation to keep the same JSON output.

Reviewed By: yangdanny97

Differential Revision: D59072598

fbshipit-source-id: 9ca44e2b0c72a1b8945c0345a5f8b6a7b7481486