improve async generator typechecking
Summary:
Currently, Pyre only infers a generator to be async if it uses `async for`, or if the the element expression is an `await` expression.

This behavior leads to some async generators being incorrectly inferred as synchronous generators, as seen in https://fb.workplace.com/groups/pyreqa/permalink/7941176282638801/

Based on the documentation, our behavior is incorrect:

> If a comprehension contains either async for clauses or await expressions or other asynchronous comprehensions it is called an asynchronous comprehension.

> Changed in version 3.11: Asynchronous comprehensions are now allowed inside comprehensions in asynchronous functions. Outer comprehensions implicitly become asynchronous.

https://docs.python.org/3/reference/expressions.html#displays-for-lists-sets-and-dictionaries

This diff extends the `await` check to also look at the condition expression. However, I think it also does not fully implement the spec since it only checks if the outermost expression is an await without also inspecting any sub-expressions.

This "implicitly making comprehensions async" behavior seems to be tricky to implement without using a visitor to check the whole expression tree. Open to clever ideas here.

Reviewed By: grievejia

Differential Revision: D59159589

fbshipit-source-id: edcdc93161209dab379b024ee3171a4ede8d1bfa