Log out the script being run at start of integration tests
Summary:
It is a bit hard to understand what is happening if either of these integration
tests fail, particularly the unsaved changes tests.

And they fail so rarely that it can actually take some time to remember where the
test module even lives. This is a silly way to waste time, the logs ought to be
obvious.

Let's add a log statement when the `run` functions kick off saying
which test
is being run (and what source file defines it).

Reviewed By: grievejia

Differential Revision: D59327379

fbshipit-source-id: f5af6138996f813e04f39b07f81e7e0844bf1a9a