Re-wire Environment to depend (mostly) on SourceCodeApi
Summary:
I want to make it possible for UnannotatedGlobalEnvironment to
be an Environment layer, which means I need the PreviousEnvironment
interface to be satisfiable by a "source code environment" that will
wrap up the source code API in an environment-functor friendly way.

To do this, I need to remove all references to UnannotatedGlobalEnvironment
from the `PreviousEnvironment` signature that the functor relies on.

This commit starts the process by moving most of the lookups to rely on
SourceCodeIncrementalApi rather than UnannotatedGlobalEnvironment. I am
not yet dealing with `CreateHandle`, but I've cut over all of the
"query" style functions.

I'll stack the create fix, which I think will require making a new interface
that wraps `SourceCodeIncrementalApi` with the "maybe ast environment" field
so that this logic no longer lives in UnannotatedGlobalEnvironment (because
functorizing requires that it live at the lowest-level previous environment,
hence *below* the lowest-level functorized environment), in a follow-up diff.

Reviewed By: grievejia

Differential Revision: D58853165

fbshipit-source-id: 4d9d945082f82b246a48099dbd372e0360aa3048